+++ api/texture.C
@@ -1,3 +1,4 @@
+
 #include <stdio.h>
 #include <stdlib.h> 
 #include <string.h>
@@ -224,11 +225,11 @@ int checkSize (int x){
 
 unsigned int texFormat;
 
-unsigned char* getRGBA (FILE *s, int size){
+unsigned char* getRGBA (FILE *s, size_t size){
     unsigned char *rgba;
     unsigned char temp;
-    int bread;
-    int i;
+    size_t bread;
+    size_t i;
     rgba=(unsigned char*)malloc(size * 4); 
     if (rgba == NULL) return 0;
     bread = fread (rgba, sizeof (unsigned char), size * 4, s); 
@@ -250,11 +251,11 @@ unsigned char* getRGBA (FILE *s, int size){
 //	getRGB
 //	Reads in RGB data for a 24bit image. 
 //	=============
-unsigned char* getRGB (FILE *s, int size){
+unsigned char* getRGB (FILE *s, size_t size){
     unsigned char *rgb;
     unsigned char temp;
-    int bread;
-    int i;
+    size_t bread;
+    size_t i;
     rgb=(unsigned char*)malloc(size * 3); 
     if (rgb == NULL) return 0;
     bread = fread (rgb, sizeof (unsigned char), size * 3, s);
@@ -276,9 +277,9 @@ unsigned char* getRGB (FILE *s, int size){
 //	getGray
 //	Gets the grayscale image data.  Used as an alpha channel.
 //	=============
-unsigned char* getGray (FILE *s, int size){
+unsigned char* getGray (FILE *s, size_t size){
     unsigned char *grayData;
-    int bread;
+    size_t bread;
     grayData=(unsigned char*)malloc (size);
     if (grayData == NULL) return 0;
     bread = fread (grayData, sizeof (unsigned char), size, s);
